+++ __tests__/01_basic_spec.js
@@ -24,7 +24,7 @@ describe('basic spec', () => {
     );
     const renderHi = () => <h1>Hi</h1>;
     const DelayedMessage = ({ delay }) => {
-      const asyncTask = useAsyncTaskTimeout(renderHi, delay);
+      const asyncTask = useAsyncTaskTimeout(renderHi, delay, [renderHi, delay]);
       useAsyncRun(asyncTask);
       const {
         pending,