+++ src/use-async-task-axios.js
@@ -1,28 +1,8 @@
-import { useRef } from 'react';
 import axios from 'axios';
 import { useAsyncTask } from './use-async-task';
 import { useAsyncRun } from './use-async-run';
 
-const shallowArrayEqual = (a1, a2) => {
-  if (a1.length !== a2.length) return false;
-  for (let i = 0; i < a1.length; i += 1) {
-    if (a1[i] !== a2[i]) return false;
-  }
-  return true;
-};
-
-// this can be too naive
-export const useMemoPrev = (create, deps) => {
-  const memoized = useRef(null);
-  const prevDeps = useRef(null);
-  if (!prevDeps.current || !shallowArrayEqual(prevDeps.current, deps)) {
-    prevDeps.current = deps;
-    memoized.current = create();
-  }
-  return memoized.current;
-};
-
-export const useAsyncTaskAxios = config => useAsyncTask(
+export const useAsyncTaskAxios = (config, deps) => useAsyncTask(
   (abortController) => {
     const source = axios.CancelToken.source();
     abortController.signal.addEventListener('abort', () => {
@@ -33,7 +13,7 @@ export const useAsyncTaskAxios = config => useAsyncTask(
       cancelToken: source.token,
     });
   },
-  [config],
+  deps,
 );
 
 export const useAxios = (...args) => {