+++ src/use-async-task-timeout.d.ts
@@ -1,9 +1,10 @@
 // eslint-disable-next-line import/named
-import { AsyncTask } from './index';
+import { AsyncTask, DependencyList } from './index';
 
 export type UseAsyncTaskTimeout = <Result>(
   func: () => Result,
   delay: number,
+  deps: DependencyList,
 ) => AsyncTask<Result>;
 
 export const useAsyncTaskTimeout: UseAsyncTaskTimeout;